+++ spec/dummy_app/app/controllers/widgets_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class WidgetsController < ApplicationController
-  def paper_trail_enabled_for_controller
+  def paper_trail_enabled_for_controller?
     request.user_agent != "Disable User-Agent"
   end