+++ CHANGELOG.md
@@ -11,7 +11,7 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- None
+- [#1524](https://github.com/paper-trail-gem/paper_trail/pull/1524) - Add rails optimistic locking support
 
 ### Fixed